Arreglado el problema de la inicializacion de firebase
