+++ firebase-storage.js
@@ -1,15 +1,7 @@
 'use strict';
-
-var admin = require('firebase-admin');
-// eslint-disable-next-line max-len
-var serviceAccount = require(process.env.FIREBASE_CREDENTIALS);
+const { admin } = require('./firebase-config');
 
 async function deleteFile(file_name){
-  admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount),
-    databaseURL: process.env.FIREBASE_DATABASE_URL,
-    storageBucket: process.env.FIREBASE_BUCKET_NAME,
-  });
 
   console.log('Firebase instance initialized');
